+++ src/e2e/e2e.constant.ts
@@ -26,7 +26,6 @@ export const AddressEventNetworks = [
   Network.DOGECOIN_TESTNET,
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -71,7 +70,6 @@ export const IncomingNativeTxNetworks = [
   Network.DOGECOIN_TESTNET,
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -116,7 +114,6 @@ export const OutgoingNativeTxNetworks = [
   Network.LITECOIN_TESTNET,
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -157,7 +154,6 @@ export const OutgoingNativeTxNetworks = [
 export const OutgoingFailedNetworks = [
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -190,7 +186,6 @@ export const OutgoingFailedNetworks = [
 export const PaidFeeNetworks = [
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -227,7 +222,7 @@ export const PaidFeeNetworks = [
 export const FungibleTxNetworks = [
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
+
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -264,7 +259,7 @@ export const FungibleTxNetworks = [
 export const NftNetworks = [
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
+
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -301,7 +296,7 @@ export const NftNetworks = [
 export const MultitokenNetworks = [
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
+
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -333,7 +328,7 @@ export const MultitokenNetworks = [
 export const FailedTxPerBlockNetworks = [
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
+
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -370,7 +365,7 @@ export const FailedTxPerBlockNetworks = [
 export const ContractAddressLogEventNetworks = [
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
+
   Network.ETHEREUM_HOODI,
   Network.POLYGON,
   Network.POLYGON_AMOY,
@@ -397,7 +392,7 @@ export const ContractAddressLogEventNetworks = [
 export const InternalTxNetworks = [
   Network.ETHEREUM,
   Network.ETHEREUM_SEPOLIA,
-  Network.ETHEREUM_HOLESKY,
+
   Network.ETHEREUM_HOODI,
   Network.CELO,
   Network.CELO_ALFAJORES,